[qsr] fix telltales align when safe window is moved to cluster window
Task-number: AUTOSUITE-1150

+++ src/neptune3-ui-qsr-cluster/safewindow.cpp
@@ -115,7 +115,7 @@ void SafeWindow::render(QPainter *painter)
 void SafeWindow::moveWindow(quint32 x, quint32 y)
 {
     setWindowState(Qt::WindowNoState);
-    setPosition(x, y);
+    setPosition(static_cast<int>(x), static_cast<int>(y));
     raise();
     requestActivate();
 }
